requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv
httpx  # для асинхронных запросов к DeepSeek
pydantic